Move Groq tool call fake into shared test helpers
