+++ tests/Helpers.php
@@ -58,6 +58,35 @@ function fakeGroqResponse(string $text = 'Hello'): PromiseInterface
     ]);
 }
 
+function fakeGroqToolCallResponse(): PromiseInterface
+{
+    return Http::response([
+        'id' => 'chatcmpl-tool-123',
+        'object' => 'chat.completion',
+        'model' => 'openai/gpt-oss-20b',
+        'choices' => [[
+            'index' => 0,
+            'message' => [
+                'role' => 'assistant',
+                'content' => null,
+                'tool_calls' => [[
+                    'id' => 'call_123',
+                    'type' => 'function',
+                    'function' => [
+                        'name' => 'FixedNumberGenerator',
+                        'arguments' => '{}',
+                    ],
+                ]],
+            ],
+            'finish_reason' => 'tool_calls',
+        ]],
+        'usage' => [
+            'prompt_tokens' => 10,
+            'completion_tokens' => 5,
+        ],
+    ]);
+}
+
 function fakeOllamaResponse(string $text = 'Hello'): PromiseInterface
 {
     return Http::response([
